pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "A server exposing tools via the Model Context Protocol (MCP)."
authors = ["Gemini"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = "^2.6.4"
mysql-connector-python = "^8.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
